Refactor InfoDialog
